+++ app/src/main/kotlin/com/maltaisn/recurpicker/demo/DateDialogFragment.kt
@@ -17,66 +17,37 @@
 package com.maltaisn.recurpicker.demo
 
 import android.annotation.SuppressLint
+import android.app.DatePickerDialog
 import android.app.Dialog
-import android.content.DialogInterface
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.widget.DatePicker
+import androidx.core.os.bundleOf
 import androidx.fragment.app.DialogFragment
-import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import java.util.Calendar
 
 /**
  * Dialog used to choose the start date.
  * This is mostly the same as [com.maltaisn.recurpicker.picker.DateDialogFragment].
  */
-internal class DateDialogFragment : DialogFragment() {
-
-    var date: Long = System.currentTimeMillis()
-
-    private var _datePicker: DatePicker? = null
-    private val datePicker get() = _datePicker!!
-
-    private val calendar = Calendar.getInstance()
+class DateDialogFragment : DialogFragment() {
 
     @SuppressLint("InflateParams")
     override fun onCreateDialog(state: Bundle?): Dialog {
-        calendar.timeInMillis = state?.getLong("date") ?: date
-        calendar.setToStartOfDay()
-
         val context = requireContext()
 
-        val view = LayoutInflater.from(context).inflate(R.layout.dialog_date, null, false)
-        _datePicker = view.findViewById(R.id.date_picker)
-        datePicker.init(calendar[Calendar.YEAR], calendar[Calendar.MONTH],
-            calendar[Calendar.DATE], null)
+        val args = requireArguments()
+        val date = args.getLong(KEY_DATE)
 
-        try {
-            // Fixes date picker showing year 1964 or 2100 initially when in spinner mode (API < 21).
-            // See https://stackoverflow.com/a/19125686
-            @Suppress("DEPRECATION")
-            val cal = datePicker.calendarView
-            if (cal != null) {
-                // Add approximatively 1 day (arbitrary).
-                @Suppress("MagicNumber")
-                cal.setDate(date + 100000000, false, true)
-                // Note that min and max date were set to a time at the start of the day earlier because
-                // date picker considers time of the day when checking if time millis is within bounds.
-                cal.setDate(date, false, true)
-            }
-        } catch (e: UnsupportedOperationException) {
-            // API >= 21, nothing to fix.
-        }
+        val calendar = Calendar.getInstance()
+        calendar.timeInMillis = date
 
-        return MaterialAlertDialogBuilder(context)
-            .setView(datePicker)
-            .setPositiveButton(android.R.string.ok) { _, _ ->
-                calendar.set(datePicker.year, datePicker.month, datePicker.dayOfMonth)
-                calendar.setToStartOfDay()
-                callback?.onDateDialogConfirmed(calendar.timeInMillis)
-            }
-            .setNegativeButton(android.R.string.cancel) { dialog, _ -> onCancel(dialog) }
-            .create()
+        // Since this is not a material component, colorAccent should be defined as either
+        // colorPrimary or colorSecondary in styles.xml. (see demo app)
+
+        return DatePickerDialog(context, { _, year, month, day ->
+            calendar.set(year, month, day)
+            calendar.setToStartOfDay()
+            callback?.onDateDialogConfirmed(calendar.timeInMillis)
+        }, calendar[Calendar.YEAR], calendar[Calendar.MONTH], calendar[Calendar.DATE])
     }
 
     private fun Calendar.setToStartOfDay() {
@@ -86,18 +57,6 @@ internal class DateDialogFragment : DialogFragment() {
         this[Calendar.MILLISECOND] = 0
     }
 
-    override fun onDismiss(dialog: DialogInterface) {
-        super.onDismiss(dialog)
-        _datePicker = null
-    }
-
-    override fun onSaveInstanceState(state: Bundle) {
-        super.onSaveInstanceState(state)
-
-        calendar.set(datePicker.year, datePicker.month, datePicker.dayOfMonth)
-        state.putLong("date", calendar.timeInMillis)
-    }
-
     private val callback: Callback?
         get() = (parentFragment as? Callback)
             ?: (targetFragment as? Callback)
@@ -105,5 +64,16 @@ internal class DateDialogFragment : DialogFragment() {
 
     interface Callback {
         fun onDateDialogConfirmed(date: Long)
+    }
+
+    companion object {
+
+        private const val KEY_DATE = "date"
+
+        fun newInstance(date: Long): DateDialogFragment {
+            val dialog = DateDialogFragment()
+            dialog.arguments = bundleOf(KEY_DATE to date)
+            return dialog
+        }
     }
 }